streamlit-app/requirements.txt
streamlit>=1.28.0
anthropic>=0.8.0
pyyaml>=6.0
pandas>=2.0.0
plotly>=5.0.0
python-dotenv>=1.0.0
